Rename datasets/datasets.py to datasets/core.py
