+++ gammapy/datasets/__init__.py
@@ -1,1 +1,1 @@
-from .datasets import *
+from .core import *
