+++ gammapy/datasets/core.py
