+++ gammapy/modeling/fit.py
@@ -2,7 +2,7 @@
 import logging
 import numpy as np
 from astropy.utils import lazyproperty
-from gammapy.datasets.datasets import Datasets
+from gammapy.datasets import Datasets
 from .iminuit import confidence_iminuit, covariance_iminuit, mncontour, optimize_iminuit
 from .scipy import confidence_scipy, optimize_scipy
 from .sherpa import optimize_sherpa
